+++ dds_web/api/marshmallows.py
@@ -1,18 +1,31 @@
+"""Marshmallow schemas used by the DDS"""
+
+# IMPORTS ################################################################################ IMPORTS #
+
+# Standard library
+
+# Installed
 import marshmallow
+
+# Own modules
+from dds_web import db
 from dds_web.database import models
+from dds_web.crypt import auth
+
+# SCHEMAS ################################################################################ SCHEMAS #
 
 
 class InviteUserSchema(marshmallow.Schema):
     """Schema for InviteUser endpoint"""
 
-    email = marshmallow.fields.Email(required=True)
+    email = marshmallow.fields.Email(required=True)  # Validator below
+    facility_name = marshmallow.fields.String(required=False)  # Validator below
     role = marshmallow.fields.String(
         required=True,
         validate=marshmallow.validate.OneOf(
             choices=["facility", "researcher"],
         ),
     )
-    facility_name = marshmallow.fields.String(required=False)
 
     @marshmallow.pre_load
     def verify_existing_facility(self, in_data, **_):
@@ -45,6 +58,7 @@ class InviteUserSchema(marshmallow.Schema):
     def make_invite(self, data, **kwargs):
         """Deserialize to an Invite object"""
 
+        # Get facility id for foreign key in invite
         facility_id = (
             (
                 models.Facility.query.filter_by(name=data.get("facility_name"))
@@ -55,6 +69,7 @@ class InviteUserSchema(marshmallow.Schema):
             else None
         )
 
+        # Create and return invite row
         return models.Invite(
             **{
                 "email": data.get("email"),
@@ -66,7 +81,9 @@ class InviteUserSchema(marshmallow.Schema):
 
 
 class NewUserSchema(marshmallow.Schema):
+    """Schema for NewUser endpoint"""
 
+    # TODO: Look through and match to db
     username = marshmallow.fields.String(required=True)
     password = marshmallow.fields.String(
         required=True, validate=marshmallow.validate.Length(max=120)
@@ -82,18 +99,32 @@ class NewUserSchema(marshmallow.Schema):
     facility_name = marshmallow.fields.String(required=False)
 
     class Meta:
+        """Exclude unknown fields e.g. csrf etc that are passed with form"""
+
         unknown = marshmallow.EXCLUDE
+
+    @marshmallow.pre_load
+    def hash_password(self, in_data, **_):
+
+        if not in_data.get("password"):
+            raise marshmallow.ValidationError("Password required to create users.")
+
+        in_data["password"] = auth.gen_argon2hash(password=in_data.get("password"))
+
+        return in_data
 
     @marshmallow.post_load
     def make_user(self, data, **kwargs):
-        """Deserialize to an Invite object"""
+        """Deserialize to an User object"""
 
+        # Get facility row
         facility = (
             (models.Facility.query.filter_by(name=data.get("facility_name")).first())
             if data.get("facility_name")
             else None
         )
 
+        # Create new user row
         new_user = models.User(
             **{
                 "username": data.get("username"),
@@ -105,10 +136,19 @@ class NewUserSchema(marshmallow.Schema):
             }
         )
 
+        # Create new email and append to user relationship
         new_email = models.Email(email=data.get("email"), primary=True, user=new_user)
         new_user.emails.append(new_email)
 
+        # Delete old invite
+        old_email = models.Invite.query.filter(models.Invite.email == new_email.email).first()
+        db.session.delete(old_email)
+
+        # Append user to facility users
         if facility:
             facility.users.append(new_user)
 
-        return new_user
+        db.session.add(new_user)
+        db.session.commit()
+
+        return new_user.username
